+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.2.0
+- First tag is now automatically trimmed when symbol copied from localization file;
+- Symbol list is now cached;
+- Big refactoring;
+
 # 1.1.0
 - Added copy_yaml_symbol_to_clipboard command (thanks to nCore);
 - Path now correctly updated after navigating to symbol;